Refactor of keras script and using model_i.py in main
